- use language class instead of hardcoded string - refactor method names to use static methods
