+++ lib/number_translator.dart
@@ -1,3 +1,4 @@
 library number_translator;
 
 export 'src/number_translator.dart';
+export 'src/languages.dart';
